requirements.txt
PyGithub
aiohttp
discord.py
audioop-lts # discord.py imports using * syntax 
python-dotenv
requests
modal
psycopg2-binary
yoyo-migrations
ruff
pre-commit
better_profanity
sqlalchemy

# api
fastapi[all] # install all to avoid random bugs
uvicorn

# docs
jinja2
requests
